Prevent errors when tapping notifications
This also makes the refresh background-resilient so it happens even when tapping notifications.

Ideally we would not fetch when notes are tapped since we are foregrounded for very little time, but in practice detecting this scenario is not possible based on the lack of concrete documentation about the ordering of the various callbacks for application state and notification tap. No solution seemed 100% so instead we absorb all errors that occur if our background time runs out.